requirements.txt
# =========================
# QML Agentic RAG
# =========================

# ---- Core orchestration ----
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0

# ---- OpenRouter (OpenAI-compatible API) ----
langchain-openai>=0.2.0
openai>=1.30.0

# ---- Embeddings / Transformers ----
sentence-transformers>=3.0.0
transformers>=4.45.0
torch>=2.2.0
accelerate>=0.33.0
sentencepiece>=0.2.0
tiktoken>=0.7.0

# ---- Vector DB ----
faiss-cpu>=1.8.0

# ---- PDF ingestion ----
pypdf>=4.0.0
pymupdf>=1.24.0

# ---- Data / utilities ----
pandas>=2.0.0
numpy>=1.26.0
tqdm>=4.66.0
python-dotenv>=1.0.0
pydantic>=2.7.0

# ---- Graph visualization ----
grandalf>=0.8
ipython>=8.0.0